// src/sharedTypes/types.ts
export type TrackType = {
  _id: number;
  name: string;
  author: string;
  release_date: Date | string;
  genre: string[];
  duration_in_seconds: number;
  album: string;
  logo: string | null;
  track_file: string;
  stared_user: [];
};

export type SelectionType = {
  _id: number;
  name: string;
  items: number[];
  owner: number[];
  __v: number;
};